requirements.txt
# Discord 봇
discord.py>=2.3.0
python-dotenv>=1.0.0

# 데이터베이스
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 유틸리티
python-dateutil>=2.8.0